Encoding predicates in function types
Summary:
**DISCLAIMER: This is work in progress. Please see limitations below.**

The goal here is to encode refining predicates of functions.

Take for example the following function that establishes the fact that its
argument is a string

  function is_string(x) {
    return typeof x === "string";
  }

One possible type we can assign is

  (x: mixed) => boolean

however, this type ignores the essence of the check performed in
`is_string`'s body. And so using this function as a conditional check

  if (is_string(x)) { ... }

does not narrow down the type for `x` in the then-branch.

A more precise description for `is_string` encodes the function's
contract in its return type:

  function is_string(x: mixed): $StrP { ... }

Here, the type `$StrP` as the return type denotes that

>  parameter `x` is a string <==> the function returns true

The scope of this mechanism is still very **limited**:

  # The only supported predicates are `$StrP`, `$NumP` and `$BoolP`.
  # Only one argument is supported, and the only context for `$*P` is the return type of a function.
  # Only functions are supported (no methods).
  # Inference of function predicates is limited to the context of If-statement conditions.
  # There is **no** support for **checking** the function's body against its specification.

So really this only works for declared functions. The immediate next
step is to infer such predicate from the function body, which would
obviate away the need for the (trickier) checking of the specification.

Reviewed By: avikchaudhuri

Differential Revision: D3588421

fbshipit-source-id: e7aee2acd0f774196e0da25c45fe8e26261e767b