fix confusing global object errors
Summary:
Function type annotations have their `this` slot open, so that they can monitor their calls and pass back the receivers of those calls as constraints on the functions/methods that they're typing.

Such function type annotations often appear in object types, and have unfortunate interactions with class method types or other kinds of bound methods, for which we considered `this` to be `mixed`. That was a mistake. The `this` of bound methods should really be `any`, because we truly don't care: `mixed` works in one direction but not in the other, and when function type annotations appear in invariant positions (like in object types) this difference becomes critical. Unfortunately the resulting errors are not only spurious but also near-impossible to debug without inside knowledge, even with traces.

Reviewed By: samwgoldman

Differential Revision: D3619932

fbshipit-source-id: c7b3f36f1c9ec5cd04233a2fade94b971b5a198f