[hack] Fix Windows build after memfd_reserve fixes
Summary:
We deleted a `memfd_reserve` call inside of `hh_collect` since it was causing trouble. However this broke the Windows build (thanks AppVeyor for noticing!). The problem is that Windows needs to reserve the memory inside of `hh_collect`.

So this is hopefully the last diff to address this. This restores the `memfd_reserve` behavior, but only for Windows. It also renames the function a little to make it clear that it has nothing to do with `memfd`.
Closes https://github.com/facebook/flow/pull/2154

Reviewed By: dlreeves

Differential Revision: D3635770

Pulled By: gabelevi

fbshipit-source-id: 5468f49ab0831f431160b28f6affca539300513e