Fill out console definition
Summary:
The motivation for this change was `s/groupCollapse/groupCollapsed`. I
noticed that there were a bunch of errors about `groupCollapsed` when testing
v0.30.0.

Since `console` is going from `any` to an object in v0.30.0, I'm a little
nervous about the errors this will cause. So I figured I might as well fill it
out a little more. I loaded
https://developer.mozilla.org/en-US/docs/Web/API/Console and added anything
listed there.

Reviewed By: samwgoldman

Differential Revision: D3636982

fbshipit-source-id: b6d42497d6baec7c034afa2d1c29344ea12d6638