Inferred function predicates
Summary:
Functions can now be declared as "predicates". The **temporary** way
to declare such a function is to prefix its name with `$pred$`. For example:

  function $pred$is_string(y) {
    return typeof y === "string";
  }

encodes the predicate of whether the argument passed is a string or not.
The body of these functions can only be of the form:

  return <exp>;

In addition the syntax for declared predicate functions changed slightly
to the following:

  declare function is_string(x: mixed): $StrP<0>;

The numeric literal in angles specifies the argument index that the
predicate refers to.

Due a complication with an orthogonal source of unsoundness, we currenly
limit support to function with a single parameter. However, the implementation
can readily support more parameters.

Reviewed By: avikchaudhuri

Differential Revision: D3635177

fbshipit-source-id: d4d3dfa9bf48f7f28b208e1348925d95dadc33f6