[hack] Introduce Decl_heap
Summary:
As part of reworking how we store class_types we need to make a distinction between the decl phases view of the shared heap and the typing phase. In particular the decl phase is the only one that should have write access to the heap. The typing phase can only read from the heap.

This will also allow us to have different views of the data available in the shared memory between the different phases.

Depends on D3538528

Reviewed By: alexchow

Differential Revision: D3539297

fbshipit-source-id: 3cdd12d81ceffa1d79be6938e579f46885f43d51