Lookup Object builtin with a reason that has a position
Summary:
The following code results in an internal assertion error when run with the
`--no-flowlib` option:

```
/* flow */
declare var o: {};
o.toString();
```

```
$ flow check --no-flowlib
add_output: no source for error: [
  {
    "kind":"infer",
    "level":"error",
    "message":[
      {
        "descr":"object type",
        "type":"Blame",
        "path":"",
        "line":0,
        "endline":0,
        "start":1,
        "end":0
      },
      {
        "descr":"Could not resolve name",
        "type":"Comment",
        "path":"",
        "line":0,
        "endline":0,
        "start":1,
        "end":0
      }
    ]
  }
]:
```

This is happening because we look up the Object builtin using the position-less
reason created for the MixedT we use to represent the Object prototype.

However, the strict reason does have a position, and is the same thing we would
have gotten had we repositioned. With this code, we get the following error:

```
test.js:4
  4: o.toString();
       ^ object type. Could not resolve name
```

Reviewed By: bhosmer

Differential Revision: D3653157

fbshipit-source-id: 96340b8c0de14b36c5aa9351c12172980c29780d