any propagation of exports
Summary:
Importing from unchecked modules that are not declared, or from modules whose
exports are `any`, used to result in "dangling" tvars. This was mostly fine
because flows from `any` to other uses were dropped anyway, but recently we have
discovered that we do need to propagating those `any`s for some parts of the
inference engine to work properly (overload checking and predicate functions
being at least two).

This diff continues that line of work. In particular it solves an issue with
predicate function calls waiting indefinitely instead of moving on when they
came from the above kinds of imports.

Differential Revision: D3655122

fbshipit-source-id: 9b797a7d95ab260f68deb3555375eaee2d198d7c