type destructors part 1: property type
Summary:
D3640942 killed a hacky way of getting the type of property 'x' of an object
type. This diff provides a more systematic replacement, beginning a line of work
on a set of type-level destructors that can introspect on types and extract
their parts (not only object types but also function types, array types, etc).

Technically it seemed reasonable to reuse the mechanism we already have in place
for destructuring: for property extraction it provides exactly the operation we
need in this diff, and it can be generalized to other cases in the future.

Reviewed By: jeffmo

Differential Revision: D3662084

fbshipit-source-id: 095942438e9f9c5acb78783ca2c8454ee9712307