[hack] Made garbage collector scale with latest heap size
Summary: The garbage collector currently will repeatedly try to collect if the memory used is greater than some constant multiple of initial memory state. With lazy decl, the initial memory state is no longer a good estimate of this, so this change allows the garbage collector to take the latest heap size after collecting as a new estimate of correct memory usage.

Reviewed By: dabek

Differential Revision: D3666178

fbshipit-source-id: d4732d6d6b0e7c3a059e858b059e1afa9bea466c