Parser support for function predicates
Summary:
Introducing the form

  declare function f(x:S,..): T checks(e);

The `checks` annotation denotes that this function will be used as
a predicate. The expression in parentheses encodes the condition
that is established by `f`.

Functions that specify a condition in their bodies omit the expression
`e` in the above declaration:

  function g(x:S,...): T checks {
    return e;
  }

This function definition states that `g` will be used as a predicate,
whose check is encoded in its body (here `e`).

For the moment predicates cannot annotate types. So the following form
is disallowed:

  var h: (x: S) => T checks(typeof x === "string") = ...;

Reviewed By: jeffmo

Differential Revision: D3668090

fbshipit-source-id: 6b01f863c420fb45b0030abad62c51210b4d9a71