Limit length of type hints
Run on this file here:
https://github.com/facebook/nuclide/blob/e679970e6a20af834e9862fb83e49a4abaf6549c/pkg/nuclide-reprint-js/lib/printers/complex/printBinaryExpression.js

`flow type-at-pos printBinaryExpression.js 43 13`

Before this change, Flow printed so much text that my tmux session was
overwhelmed for a couple minutes. When Nuclide called it, Atom would crash. Now,
it still takes 15 seconds to run but the output is smaller and Atom doesn't
crash when Nuclide calls `type-at-pos`. Flow features are unavailable while the
type is being computed, but that's better than the entire editor crashing.