promise resolve void fix
Summary:
The declared type of `Promise.resolve` takes an optional parameter. This probably
predates a change in Flow's behavior where a missing arg in a call is the same
as passing `undefined`. In particular, making the parameter of `Promise.resolve`
required has exactly the desired behavior: when `undefined` is passed it returns
`Promise<void>`.

Reviewed By: samwgoldman

Differential Revision: D3695806

fbshipit-source-id: d960495e5f198891dbabdef632505e0bfc0e83f9