async return void fix
Summary:
Previously, `async` functions with return values would call Promise.resolve on
those values to create the desired return type.

Unfortunately, `async` functions that with undefined return values would end up
returning Promise<T> for some unconstrained T, because of an idiosyncrasy in the
declared type of Promise.resolve: that method takes an optional parameter, so
passing undefined values to it makes it completely bypass the intended type
constraints.

This diff instead calls $await on return values followed by explicit wrapping
Promise to derive the return type of async functions. This fixes the immediate
issue. It is also arguably more correct to use $await instead of Promise.resolve
to derive the return type, since async and await should behave symmetrically
when the global Promise is overridden.

Separately, the declared type of Promise.resolve seems to be buggy and will be
fixed in a follow-up diff.

Reviewed By: samwgoldman

Differential Revision: D3695781

fbshipit-source-id: 184961a00a55b75afd1600cf959b6e07cf62a290