non-null-fix and type destructors rewrite
Summary:
A little while back `$NonNull<T>` was added as a type operator using `*` under the
hood to propagate constraints. Unfortunately it is a bit fragile.

While it's cool that we can express things like `$NonNull<T>` (and we do some of
that in our React model), the general approach has some serious limitations.
(1) `*` can be pin types down in undesirable ways, as we've seen with
`ReactElement<*>`.
(2) Constraints can lead to underspecification, as we've seen
in the definition of `ReactClass<T>`.

Meanwhile we've begun a line of work to add type destructors (starting with
`$PropertyType`). This diff adds another one, `$NonNullVoidType`. It basically takes
over the job of `$NonNull<T>`, which can now be removed.

Overall this diff makes the feature far more stable than possible with `*`
alone. We can of course generalize the process of making destructors and put it
back into the type system, but that's for later.

Fixes https://github.com/facebook/flow/issues/2199

Reviewed By: samwgoldman

Differential Revision: D3680253

fbshipit-source-id: 6c15ed3608e48604acda24975945f8a795b7b9ba