Add method signature annotation for class example
Summary:
Addresses the confusion created by the current `React.Component` subclass example and described in #1694

Could also create a type for `state` as well to keep it consistent. If this isn’t desired and it’s preferred to keep the example with the inline type declaration for the `props` instance property, I can instead add a note after the example warning users that any exported component will require this kind of annotation.

When I was trying to figure this out in my own project, I had hoped that there would be a way to reference the inline `props` annotation in the `constructor` method signature, but couldn’t figure it out. If that’s possible in a nice way, I think changing this example to use that technique would be best.
Closes https://github.com/facebook/flow/pull/2213

Reviewed By: samwgoldman

Differential Revision: D3694093

Pulled By: gabelevi

fbshipit-source-id: db9c1206c3312887bf6888816c98165e936d925a