[hack] Move Sys_utils.cat to a separate Disk module and make it injectable, create a test version
Summary:
Disk access is an operation on a global state, and threading something representing a file source to all the callsites of Sys_utils.cat would be a significant refactoring.

Instead of doing this, I make the injectable version of this function also operate on global state - but instead of disk contents, it's state that we can control through TestDisk.set function.

Reviewed By: dlreeves

Differential Revision: D3700296

fbshipit-source-id: 80bf26141b19729343f7a311708bacfa73982356