Consider that exact yet unsealed objects may be wider than they seem
Summary:
All object literals are exact, including empty literals which are also unsealed
in the file where they are created. It turns out that an unsealed object, much
like inexact objects, can be wider than they seem at any given time, not through
subtyping but rather because properties can be added later.

This diff changes the behavior of `ObjT -> ExactT` and the property existence
predicate, which both have special logic for exact object literals which does
not apply to unsealed objects for this reason.

Reviewed By: avikchaudhuri

Differential Revision: D3708396

fbshipit-source-id: 412f8ec3768c55822aaae4dac12f720b31bc86c4