Add catch pattern to type table
Summary:
Previously the pattern in a catch clause was missing from the type table. This
commit adds that tvar so it can be found when looked up.

Note that the tvar will be open in practice, because we do not flow thrown
values into the catch. This is hard to do in general, because the throw can
happen in a function call (we would need to store thrown types in the function
type). Also, the value is not always an Error; anything can be thrown.

Still, this is notably missing and easy to add. The necessary type inference
support would be interesting to do, but I'm leaving it for future work.

Reviewed By: bhosmer

Differential Revision: D3743667

fbshipit-source-id: 8d650f54c037e48210e063f448945050fe0a3c15