[hack] Skip logging in worker for early failures
Summary:
If a worker fails then it calls EventLogger.worker_exception with the exception
string. But if the worker fails before initializing logging, or doesn't
initialize it at all, then it couldn't log the exception.  This was causing us
to fail during logging, hiding the backtrace. So let's wrap it with
log_if_initialized().

Reviewed By: hermanventer

Differential Revision: D3705795

fbshipit-source-id: d0baf767c11398d7485d9089b7f03621909b9d66