[hack] Move some types from files in "server" to files in "server_services" TARGET
Summary: Most of the dependencies on "server" target are through using types defined there. In order to prepare for {D3724557}, I'm moving some of them away from this TARGET. This is in line with original design mentioned in D3724524 - notice how the result type for AUTOCOMPLETE is already in AutocompleteService, not in ServerAutoComplete.

Reviewed By: jamesjwu

Differential Revision: D3724547

fbshipit-source-id: edb6cf3756eaf6094d03be2c8c962487603fbce6