Predicate-abstract functions
Summary:
**Disclaimer: The syntax here is temporary, as this is work in progress**

Predicated functions can now be passed as arguments to other functions
and their checking "effect" can be used to refine the output.

Take for example the function `refine`:

  declare function refine<T>(v: T, f: Function): T';

This function takes as arguments a value `v` of type T and a function
`f` that we intend to use as a predicate. In particular, we test `v`
by passing it to `f` and if the result is truthy then we return the
narrowed type T' <: T that the condition of `f` implies otherwise
throw an error. For example we could pass `is_string` as `f` and a
"mixed" value as `v` and get a string as the result.

We can now write this intuition in the form of a signature for `refine`:

  declare function refine<T, P: $Pred<1>>(v: T, f: P): $Refine<T,P,0>;

This type says that `f` needs to be a predicated function that takes (and
refines) one argument. This is what the bound to `P`, nameley `$Pref<1>`
means. The return type `$Refine<T,P,0>` denotes that the type of the
result is going to be `T` refined by the predicate that function `P`
guarantees for its 0-th parameter. So, if we pass `is_string` we are
essentially narrowing `T` down by the `StrP` predicate.

A more interesting example is that of the following array-filter

  declare function filter<T, P: $Pred<1>>(v: Array<T>, f: P): Array<$Refine<T,P,0>>;
  declare var a: Array<mixed>;
  (my_filter(a, is_string): Array<string>);

that narrows down the type of the contents of an array.

Sanity checks have been added for checking the numeric inputs above.

Reviewed By: avikchaudhuri

Differential Revision: D3734638

fbshipit-source-id: 38aa937910c74a77d5288869ed0ad2a50bfd3411