fix getdef of overrides
Summary:
Jump to definition for a call to an overridden method in a subclass was pointing
to the method in the superclass instead. Turns out it was a matter of doing a
Map.union in the wrong order while building the map of properties and types
visible to the subclass.

This diff builds a better API for the job, to try preventing future occurrences
of this problem. (It has bit us in the past in some other context, too.)

This problem might have also arised in autocomplete, will add tests in a
follow-up diff.

Reviewed By: samwgoldman

Differential Revision: D3781312

fbshipit-source-id: c1ebaa94cda18894a1c9fef3462e7202dae84257