autocomplete override test
Summary:
Confirmed that the recent fix for get-def on overrides (D3781312) also
applies to autocomplete, so a regression test is warranted.

Namely, an autocomplete on `this.` in a subclass that overrides a method would
return the type of the method in the superclass, which is not necessarily the
same as the type of the method in the subclass (return subtyping being only one
way they could be different).

Reviewed By: samwgoldman

Differential Revision: D3781628

fbshipit-source-id: 4513d930118fca23a6a7b8d933c4ed0fd8b11dab