minor rationalization of module export / import
Summary:
Module export/import is a pretty complicated subsystem of Flow. This diff adds
some commentary around how ModuleT is used to describe both a growing module to
be exported as well as a fully formed imported module, and what the various
rules that hit ModuleT in the two contexts should expect to handle.

Doing this analysis uncovers a small incompleteness in ExportStarFromT
rules, which this diff also fixes. Renamed to be clearer.

Finally, this diff also adds a TODO on the current handling of `import type`,
which has a special case that was recently removed in the related type aliasing
rules.

Reviewed By: gabelevi

Differential Revision: D3776755

fbshipit-source-id: e2891bcd63371a946489f2a86ac5c040bf73a94e