[hack] Typed AST via typing hooks - TAST is a map
Summary:
Produce a Typed AST which is just the NAST along with a map for the types.

Here, the TAST types are slightly different from locl ty, with (for example)
Tvar removed and with null types. It's unclear to me what these will
eventually look like when fully flushed out (hopefully merge it into one of the
existing type systems), but keeping it separate to play fast-and-loose for now.

Reviewed By: dlreeves

Differential Revision: D3744698

fbshipit-source-id: 578262f87124bef87b2c8afb3b911052dd6d73fa