graphml output for Context.graph of checked modules
Summary:
`flow check-contents --graphml < foo.js` writes the type graph of the checked contents (after local inference and before merge) to `contents.graphml`. If a filename is given, output is written to `<filename>.graphml`.

Graphml can be loaded into a tool like yEd and examined for interesting insights. :)

Note: output contains no layout info, so a tool is necessary to put the
graph in readable form. E.g. in [yEd](https://www.yworks.com/products/yed/download),
* load the graphml file
* Tools > Fit Node To Label
* Layout > One-Click Layout (or whatever)

```
/* flow */

function foo(x: ?number): string {
  if (x) {
    return x;
  }
  return "default string";
}
```

{F63132437}

The output tries to be as unopinionated as possible, but it makes a
couple of choices:

* tvars are drawn connected only to their immediate concrete bounds, not to the transitive closure that they actually hold in their lower and upper maps.
* for values within a type that aren't actually other types, we're relying on `Debug_js.dump_t` to provide the essentials for node labels. Some types don't print much in `dump_t` - any upgrades there will carry through to here.

Reviewed By: gabelevi

Differential Revision: D3786537

fbshipit-source-id: eeff2859977671cdbb868b4bc3ca9f4378429503