./tool babel
Summary:
One of the most annoying things about `./tool` is the start up time of
`require('babel-register')`. This introduces `./tool babel` which runs babel on
`tool` and the tests.

There are probably a bunch of workflow things that can be built on top of this.
Like maybe something that transforms the tests as you edit them.

Reviewed By: jeffmo

Differential Revision: D3441404

fbshipit-source-id: a5d8e2c79bf8a5c18f8aea40ea6db709a36c2922