[hack] Move Diagnostic_subscription from utils to server, refactor it
Summary:
In the next diff it's going to start depending on error list, which would create a cyclic dependency. It's used only inside of server, so it's questionable whether it actually is an "util".

It's also hiding an option inside, but I think the option should be explicit, so the callsites can leverage it for more typesafe code instead of remembering that it's safe to call "get_id"" only after checking "is_empty" etc.

Reviewed By: oulgen

Differential Revision: D3769189

fbshipit-source-id: 236aa97cc8ebaac13a38ff22d9482d91800712b5