Add `--out-dir` flag and accept a directory
Summary:
This begins the process of making the command mirror the [Babel CLI API](https://babeljs.io/docs/usage/cli/).

First up: This diff makes it possible for the command to take either a file or a directory and an optional `--out-dir` flag.

**If a file is passed,** the `.flow` file is generated and printed to stdout if no `--out-dir` flag is specified. If a `--out-dir` flag is specified, the file will be written to `<out-dir>/<filename>.flow`.

**If a directory is passed,** we require that `--out-dir` is specified, find all files (equivalent to `flow ls`) under the specified directory, generate `.flow` files for each of them, and emit them into `<out-dir>/` with the same directory structure as the original files.

thejameskyle: It would also be good to get your eyes on this as we aim to mirror the babel-cli interface as this command grows in capability.

Reviewed By: gabelevi

Differential Revision: D3806342

fbshipit-source-id: 1fee564c1c3d4e0e8209a88606e3f72736e82c5a