move resource file typechecking to merge phase
Summary:
A recent change to typecheck resource files caused a large perf regression in
the merge phase.

This diff eliminates that perf regression by generalizing the trick we use for
unchecked files to manage typechecking of resource files. The crucial
observation is that exported types for resource files don't depend on their
contents, just their extension (and therefore can be generated on the fly rather
than written/read to/from contexts living in a shared heap).

Reviewed By: bhosmer

Differential Revision: D3809314

fbshipit-source-id: 92c11e80d78917d60183af9396c8812326adddcd