Fix ./tool add-comment to ignore comments in the AST
Summary:
`./tool add-comment` parses a file to an AST to figure out where to
add the comment. It tries to find the right place in the AST to add the comment
by looking for the part of the AST that falls within the error location.

Unfortunately, the algorithm was also looking at the `comments` node of the
AST. So if a comment fell within the error location, `./tool add-comment` might
not notice whether the error fell within JSX, templates, etc.

Reviewed By: jeffmo

Differential Revision: D3821808

fbshipit-source-id: 58c521e7990a3af2ccd0e486b9aed6ccb7ff619e