Do not store package.json keys with quotes
Summary:
Rather than storing each key with the surrounding quotes, and adding them back whenever we query for a key, just strip them to begin with.

Ran unit tests, and also verified that they failed with a bunch of missing package errors with only the change on line 167 and nothing else.

This will make some of the logic in #2276 a bit simpler.
Closes https://github.com/facebook/flow/pull/2413

Reviewed By: gabelevi

Differential Revision: D3829495

Pulled By: nmote

fbshipit-source-id: 3e2c8d0de46d6a7f80f219b719c765bc6826a592