replace node after emitting goto constraints
Summary:
We chain one tvar to another during unification in two steps: we replace t1
with a Goto node pointing at t2 and we emit constraints to connect t1's lowers
to t2's uppers and vice versa.

Previously we replaced the node first then emitted the constraints, but this
breaks in the presence of UnifyT. If t2's uppers contains a UnifyT that refers
back to t1, we would erroneously pick up t2's bounds during the nested
unification instead of t1's.

Reviewed By: bhosmer

Differential Revision: D3802871

fbshipit-source-id: 82fecffda563499581bb3e559c0d1f5778736483