Make Function.prototype and Object.prototype lookups consistent
Summary:
Because builtins are merged in after local inference, we use a hack whereby a
strict lookup that hits Object.prototype first checks whether the property name
matches a static list of known Object.prototype members.

This diff refines that hack and applies it to Function.prototype lookups as
well.

Before this diff, we only looked up Object.prototype from builtins for strict
lookups, but we actually need to for non-strict lookup as well. For example, a
class that overrides `toString` or `valueOf` should still be consistent with
`Object`, same as any two user-defined classes.

Furthermore, we would fire the NonstrictReturning logic in cases where the
property would be found if we tried to look at Object.prototype. For example,
`o.toString && o.toString()`.

Reviewed By: bhosmer

Differential Revision: D3760168

fbshipit-source-id: d9762c87f0899490cb12409b0ab0ecb35cae1ced