Avoid killing Flow server on package.json changes
Summary:
Flow only cares about the `main` and `name` fields in `package.json`. However if any `package.json` changes at all, the Flow server dies. This is pretty frustrating and leads to a lot of lost time for Nuclide devs (since we have `package.json` files all over the place). This change makes it so the Flow server only dies if a relevant change is detected.

This works (I've tested it manually), but I'm not super confident that it's the right way to do it:

* I have to duplicate the options computation in order to call `reparse`. Is there somewhere a utility could live that would be like `reparse` except it grabs these options itself? Would that belong in `Parsing_service_js` or do you want to keep that from taking a dependency on `Options`?
* I'm not sure what the implications are of reparsing a file. Since it modifies global state it's hard to know. Is that a safe operation here?

Once we get the implementation strategy sorted out I'll look into adding automated tests.

CC zertosh
Closes https://github.com/facebook/flow/pull/2276

Reviewed By: samwgoldman

Differential Revision: D3837003

Pulled By: nmote

fbshipit-source-id: a2fcb2a7167a9e387bcf419c0e62b70134cedc8d