[Easy] Don't print traversal warnings for ignored files
Summary:
This is primarily affecting Windows users, who ignore
`.*/node_modules/.*` but get lots of warnings when their `node_modules/` dir
has a ton of overly long paths (which are banned in Windows).

We can't really check includes and ignores for directories, since they are both
designed to be checked against files. However, we can do a best effort, which
works well enough when people ignore path/to/dir/.*

Fixes #2187

Reviewed By: avikchaudhuri

Differential Revision: D3804714

fbshipit-source-id: ef95f92af8281103b92106436fd333c4d5002132