[hack] Pass typechecker options to parsing/elaboration
Summary:
In order for us to implement namespace aliasing, we need a way to pass this information from hhconfig to parsing. To do so, I chose typechecker options. This diff refactors some parts of the code in order to enable this passing.

This change needs to be done during the parsing phase so that IDE support still works as expected, i.e. auto completion.

Reviewed By: jopamer

Differential Revision: D3800689

fbshipit-source-id: 283953b47e7c087f03135509a288669e174e43ff