[hack] Support auto namespace aliasing through hhconfig
Summary:
This diff creates the interface and the functionality to enable auto namespace aliasing by defining a variable in .hhconfig
This namespace aliasing can somewhat be considered renaming since we no longer allow the source namespace to be used by itself.
This process is done at the parsing phase so that IDE support for aliasing still exists.

Reviewed By: jopamer

Differential Revision: D3808701

fbshipit-source-id: f3fe4bbbc31a65a718fc7660dfab4fea13cfed38