fix filtering of .flowconfig libs section
Summary:
if you add a file to `[libs]` with an extension that isn't in `module.file_exts` (.js and .json by default), that file is currently silently ignored because we pass the entries through the file finder to handle directories and symlinks, but it also filters.

this diff makes it so that a file in `[libs]` is never filtered out, but a directory in `[libs]` will only include subfiles that match the file extension filter.

Fixes https://github.com/facebook/flow/issues/181

Reviewed By: gabelevi

Differential Revision: D3878614

fbshipit-source-id: 600ad37cd02036bc4e1a70490f39a1fdac192fba