[hack] Revive removed class elements in ServerIdeUtils.declare_and_check
Summary:
`ServerIdeUtils` works by parsing the contents of a buffer and determining the top level elements by parsing it. These top level elements are then moved temporarily to the old heap. We then populate the new heap by parsing, naming, and declaring the contents. Finally after doing a type check we revive the previous state by moving the top level elements back to the new heap.

This whole process seems correct, but it actually leaves the world in an inconsistent state. This is due to us folding class elements of a parent class into it's children. Take for instance the following project containing two files.

```
<?hh // strict
class Parent {
  public function meth(): void {}
  public function test(Child $child): void {
    $child->meth();
  }
}
```

```
class Child extends Parent {}
```

We have a class `Parent` that is inherited by a class `Child`. `Parent` invokes the `Parent::meth()` method on an instance of `Child`. `Child` will have  amethod table containing the methods it inherited from `Parent`, i.e. `{ 'meth' => 'Parent::meth', 'test' => 'Parent::test' }`.

Now consider what happens when the content passed to `ServerIdeUtils.declare_and_check` is the following:

```
<?hh // strict
class Parent {
  public function test(Child $child): void {
    $child->meth();
  }
}
```

Here the `Parent::meth` method is removed. Because it is removed, `Child` needs to be re-declared but we do not do this in `ServerIdeUtils` (do not know if it is by design or a bug). The state of the world is now in an inconsistent state.

Before {D3564703} we would still have the old type of `Parent::meth` folded directly in `Child`'s declaration, however afterwards we will try to look up the type of `Parent::meth` in the heap, but it won't be there. This leads to the `Not_found` exception being thrown and the server crashing.

The easiest way to fix this is to make sure any class elements that are removed in the contents are revived from the old heap before type checking. That way if there are any child classes that still refer to the non-existent member they still have a type to refer to. This matches the inconsistent state of the world we had before {D3564703} so it seems like a sensible thing to do for now. I do think we need to rethink how some of these IDE commands work, and if it matters that we are giving results based on an inconsistent state of the world.

Reviewed By: dabek

Differential Revision: D3877141

fbshipit-source-id: 4a427ff9087f688bc3ec61c707e367cd8089c5cd