Add tab to show JSON on /try
Summary:
added tabs to the result pane on /try, to toggle between formatted errors and JSON.

I also considered just adding a "[ ] JSON" checkbox but thought it looked weird by itself.

I now construct most of the dom nodes in JS, so that it's easier in the future to drop these editors on a page without a lot of boilerplate... just a div for the editor and a div for the results, and a call to `tryFlow.createEditor()`.

Reviewed By: avikchaudhuri

Differential Revision: D3887925

fbshipit-source-id: de50558c872e6d5200654cacad1492889d27fdd4