Do not die if a libfile is touched but not changed
Summary:
Previously `touch libfile` or `flow force-recheck libfile` would cause the Flow server to die. This was particularly unfortunate since Nuclide calls `force-recheck` whenever you open a file. So simply opening a libfile in Nuclide would cause the Flow server to die.

Now, it only dies if the AST for the libfile changes. This is similar to the logic in a50968e2df6fd74bcacd800b8411a21e131559e7 which, for ordinary files, only rechecks if the AST has changed.

One notable change that this introduces is that we will now store ASTs for all libfiles, which will come at the cost of some memory. I suspect it won't be enormous since in a typical project, ordinary files far outnumber lib files. But I haven't confirmed this empirically.
Closes https://github.com/facebook/flow/pull/2474

Reviewed By: avikchaudhuri

Differential Revision: D3870402

Pulled By: nmote

fbshipit-source-id: 228a4a858b130ecf21b951a7a7f7e2d6662e5eef