Don't insist that optional props exist in quick_error_fun_as_obj
Summary:
This function is designed to eliminate union cases eagerly, in order to make
progress during speculative match where we would otherwise need to wait for
libs to be merged.

Before this diff, the functionality did not match the fall-through case for
optional properties. We would get an error indicating that an optional property
was missing. After this diff, there is no error, as expected.

Reviewed By: avikchaudhuri

Differential Revision: D3853159

fbshipit-source-id: f50bddcc8b104110f9402d8b286f58bc18f53b17