[hack] Ensure ServerIdeUtils.declare_and_recheck changes are undone
Summary:
`ServerIdeUtils.declare_and_recheck` needs to leave the state of the world
exactly the same as it was before. {D3877141} attempted to fix this for one
particular case, but broke other cases in the process. I added more thorough
testing to see how `ServerIdeUtils.declare_and_recheck` behaves when
declarations are added/removed to ensure the invariant is true. This contains
the following fixes.

* Revert the changes that made `SharedMem.NoCache.revive_batch` idempotent. It turns out we need its old behavior that removes items from the New heap if they aren't present in the Old heap
* Properly cleanup HH_FIXMES. These weren't being properly reverted
* Fix clean of class elements. Since we revive removed members early, when we revive the remaining members it should be done before reviving the Decl_heap.Classes

Reviewed By: dabek

Differential Revision: D3900513

fbshipit-source-id: c2295b66e0c50bdb6cb28c66034289d4286c05fe