don't allow mixed to be added
Summary:
it was unsound to allow `mixed` to be added. for example:

```
let x = {};
x + ''; // error
(x: mixed) + ''; // ok???
```

this also applies to template literals. I changed it from modeling templates as a series of additions, to coercing each expression to string. Couldn't decide whether to use a special form of UseT, or make a new use type.

Reviewed By: jeffmo

Differential Revision: D3893837

fbshipit-source-id: c34d5cc05b4d9dec666d20ab8a8dda222e26c748