[hack] New option --all-errors on hh_single_type_check
Summary:
As it says on the tin. Currently `hh_single_type_check` displays only the first error
that it finds in the file. With the `--all-errors` option, it will show all errors,
using indentation to make the separation of the errors apparent. This makes it easier
to see how errors appear in hh_client or the IDE.

Reviewed By: oulgen

Differential Revision: D3907430

fbshipit-source-id: fe819e96385663622fd2f7d3904929f9f0ee4746