fix race condition with server restarts
Summary:
the `package_json_changes` test was racy. adding a file that should kill the server might not have been noticed by the server by the time the next action runs.

added a new action, `waitForServerToDie(timeout)`, which just sleeps until the server dies or the timeout occurs.

added a new assertion, `serverRunning(isRunning)`, which confirms that the same PID is still running (or not running).
Closes https://github.com/facebook/flow/pull/2512

Reviewed By: gabelevi

Differential Revision: D3909447

Pulled By: mroch

fbshipit-source-id: 5264d4b3169bc7c01e528db15a5c726fde560353